Starting to add some tests...
